backend_rest/pyproject.toml
[project]
name = "backend-rest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.5.0",
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
]


[tool.ruff]

line-length = 120

[tool.ruff.lint]
select = ["I"]

[tool.uv]
dev-dependencies = [
    "django-extensions>=3.2.3",
]
